Extract example into its own directory
