+++ example/public/manifest.json
