+++ example/public/robots.txt
