+++ example/src/App.tsx
@@ -15,7 +15,7 @@ import {
   TokenListContainer,
   TokenListProvider,
 } from "@solana/spl-token-registry";
-import Swap from "./swap";
+import Swap from "@project-serum/swap-ui";
 import "./App.css";
 
 // App illustrating the use of the Swap component.
