+++ example/tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "extends": "./tsconfig.paths.json",
   "compilerOptions": {
     "target": "es5",
     "lib": [
@@ -17,14 +18,14 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": false,
+    "noEmit": true,
     "jsx": "react-jsx",
-    "typeRoots": ["types/", "node_modules/@types"],
-    "outDir": "dist",
-    "declaration": true,
-    "declarationDir": "dist"
+    "typeRoots": [
+      "../types/",
+      "node_modules/@types"
+    ]
   },
   "include": [
-    "src/swap/"
+    "src"
   ]
 }
