+++ example/tsconfig.paths.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "paths": {
+      "@project-serum/swap-ui": ["../src/"]
+    }
+  }
+}
