+++ package.json
@@ -1,14 +1,14 @@
 {
   "name": "@project-serum/swap-ui",
-  "version": "0.1.0-alpha.17",
+  "version": "0.1.0-alpha.18",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
-  "homepage": ".",
+  "homepage": "https://github.com/project-serum/swap-ui",
+  "license": "Apache-2.0",
   "dependencies": {
     "@project-serum/serum": "^0.13.34",
     "@project-serum/swap": "^0.1.0-alpha.14",
-    "@solana/spl-token": "^0.1.4",
-    "react-async-hook": "^3.6.2"
+    "@solana/spl-token": "^0.1.4"
   },
   "peerDependencies": {
     "@material-ui/core": "^4.11.4",
@@ -17,17 +17,14 @@
     "@project-serum/anchor": "^0.6.0-beta.1",
     "@solana/spl-token-registry": "^0.2.86",
     "@solana/web3.js": "^1.10.1",
-    "react": "^17.0.2"
+    "material-ui-popup-state": "^1.8.3",
+    "react": "^17.0.2",
+    "react-async-hook": "^3.6.2"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "prettier src/** --check",
-    "lint:fix": "prettier src/** -w",
-    "deploy": "gh-pages -d build/",
-    "build:npm": "rm -rf dist && tsc --build tsconfig.npm.json"
+    "lint": "prettier src/** example/src/** --check",
+    "lint:fix": "prettier src/** example/src/** -w",
+    "build": "rm -rf dist && tsc --build tsconfig.json"
   },
   "eslintConfig": {
     "extends": [
@@ -35,24 +32,15 @@
       "react-app/jest"
     ]
   },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
   "devDependencies": {
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.58",
     "@project-serum/anchor": "^0.6.0-beta.2",
+    "@project-serum/serum": "^0.13.34",
     "@project-serum/sol-wallet-adapter": "^0.2.0",
+    "@project-serum/swap": "^0.1.0-alpha.14",
+    "@solana/spl-token": "^0.1.4",
     "@solana/spl-token-registry": "^0.2.86",
     "@solana/web3.js": "^1.10.1",
     "@testing-library/jest-dom": "^5.11.4",
@@ -69,6 +57,9 @@
     "notistack": "^1.0.7",
     "prettier": "^2.3.0",
     "react": "^17.0.2",
+    "react-app-rewire-alias": "^1.0.3",
+    "react-app-rewired": "^2.1.8",
+    "react-async-hook": "^3.6.2",
     "react-dom": "^17.0.2",
     "react-scripts": "4.0.3",
     "typescript": "^4.1.2",
