+++ src/components/Info.tsx
