+++ tsconfig.json
@@ -17,9 +17,12 @@
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
-    "noEmit": true,
+    "noEmit": false,
     "jsx": "react-jsx",
-    "typeRoots": ["types/", "node_modules/@types"]
+    "typeRoots": ["types/", "node_modules/@types"],
+    "outDir": "dist",
+    "declaration": true,
+    "declarationDir": "dist"
   },
   "include": [
     "src"
